# back-end/src/main/webapp/resources/swagger.yaml
swagger: '2.0'
info:
  description: >-
    iCAN! API  --- Важно --> Всё API, кроме команд login и registration,
    работает только с передачей в headers токена сессии с именем USER_TOKEN. Для
    ответов с типом ...List в ответ идет сразу массив без имени! Постраничный
    вывод - это имена типов заканчиваются на "Paged"
  version: '1.0'
  title: Backend Rest Api documentation
  contact:
    name: contact@email
  license:
    name: Swagger Editor
    url: 'http://editor.swagger.io/#/'
host: 'localhost:8080'
basePath: /
...
paths:
  '/api/file/upload/{profileId}':
    post:
      tags:
        - files
      summary: загрузка файла на сервер и привязка его как лого для профайла
      operationId: file-upload
      parameters:
        - name: profileId
          type: integer
          in: path
          required: true
        - name: file
          type: string
          in: query
          required: true
          description: бинарный файл как параметр multipart/form-data
      responses:
        '200':
          description: ok
  /api/file/download:
    get:
      tags:
        - files
      summary: загрузка файла с сервера
      operationId: file-download
      parameters:
        - name: filename
          type: string
          in: query
          required: true
          description: имя файла - будет поиск в заданной папке (проперти filestorage.dir)
      responses:
        '200':
          description: ok
...